requirements.txt
Faker==6.1.1
mysql-connector-python==8.0.23
protobuf==3.14.0
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
python-dotenv==0.15
